+++ cli/src/dag/mod.rs
@@ -73,13 +73,6 @@ pub(crate) enum DagCommand {
             value_name = "DATA"
         )]
         input_json: serde_json::Value,
-        #[arg(
-            long = "encrypt",
-            help = "Which entry ports to encrypt in the format `<vertex_name1>.<port_name1> <vertex_name2>.<port_name2>`.",
-            num_args = 0..,
-            value_name = "VERTEX.PORT"
-        )]
-        encrypt: Vec<String>,
         /// Whether to inspect the DAG execution process.
         #[arg(
             long = "inspect",
@@ -131,7 +124,6 @@ pub(crate) async fn handle(command: DagCommand) -> AnyResult<(), NexusCliError> 
             dag_id,
             entry_group,
             input_json,
-            encrypt,
             inspect,
             gas,
         } => {
@@ -142,7 +134,6 @@ pub(crate) async fn handle(command: DagCommand) -> AnyResult<(), NexusCliError> 
                 dag_id,
                 entry_group,
                 input_json,
-                encrypt,
                 inspect,
                 gas.sui_gas_coin,
                 gas.sui_gas_budget,